Combine Sample Coelutions with Peaks

For a single sample, from coelution mz pairs (essentially edges) infer the peaks [mz, scans] then join these coelution peaks with peaks in the clamr_db.

combine_sample_coelutions_w_peaks(sample_peaks, sample_coelutions, MS1tol)

Arguments

sample_peaks

peaks from a single sample from .mzrollDB

sample_coelutions

coelutions from a single sample from .mzrollDB

MS1tol

from build_clamr_config or format_mass_accuracy_input

Details

The edge-list defines the connection between pairs of mzs over a scan interval. In order to build a graph from these connections, we need to be able to define repeated observation of the same node in terms of a common mz and scan.