Identify Coelutions

Label common mass differences using either a library of commonly seen adducts and isotopologues or optimize chemical composition as a mixed integer linear programming problem.

identify_coelutions(
  coelutions,
  type = "denovo",
  coelution_n_cutoff = 0L,
  quietly = FALSE,
  coelution_library = NULL,
  amu_tolerance = NULL,
  isotope_summary = NULL,
  ...
)

Arguments

coelutions

A list containing coelutions and coelution_mz_diffs of the coelution class produced by find_common_coelutions

type

library: match mzs to a list of knowns based on a specified mass accuracy tolerance -- required objects coelution_library, amu_mass_accuracy_tolerance

denovo: match each mz to linear combinations of isotopes -- required objects: isotope_summary

coelution_n_cutoff

Minimum membership size of a mz difference group to be tested for IDs

quietly

suppress print statements

coelution_library

a data.frame containing the following required variables

mz: numeric mz differences in amu

name: character

type: the type of coelution event -- e.g., isotopologue, adduct, distinct_compound

amu_tolerance

search tolerance for matches in amu -- will report the highest priority match and otherwise the closest match

isotope_summary

isotopes that the unknown specie could be constructed with:

label

amu

count_lb (optional)

count_ub (optional)

prior_weight (optional)

If penalization = "prior", a variable named "prior" must be specified (larger weights are bad)".

...

additional arguments to pass to coelution identification