MS2 Driven Aligner Plotting

MS2 Driven Aligner Plotting

ms2_driven_aligner_plotting(
  rt_plot_data,
  print_plots,
  show_specific_samples = NULL
)

Arguments

rt_plot_data

plot_data produced using ms2_driven_aligner.

print_plots

If FALSE then return individual plots in a list; if TRUE then print a plot and silently return 1.

show_specific_samples

A vector of sampleIds or if NULL a set of sampleIds will be randomely sampled. If sampleIds are named then they will be used for plots.

Details

1: dendrogram comparing samples based on shared MS2 manhattan distance

2: tile plot comparing samples based on shared MS2 counts

3: visualize RT deviation trend for individual samples

4: compare the RT of MS/MS2 matched samples for pairs of samples

5: consensus ppm drift of shown samples

6: mass accuracy deviations of individual samples