Reextracted Peak Intensity Set

Produce a table of reextracted peak intensities based on a short list of curated peak groups

reextracted_compound_intensities(
  curated_mzrolldb,
  full_mzrolldb,
  sample_folder,
  min_mz_delta = 0.01,
  min_rt_delta = 0.1,
  verbose = TRUE,
  debug = FALSE
)

Arguments

curated_mzrolldb

mzrollDB file containing manually verified peakgroups. peaks and peakgroups from this table are used in this method.

full_mzrolldb

mzrollDB file containing original rt_update_key RT alignment table. This table is used for transforming sample_rt <--> aligned_rt

sample_folder

folder containing all raw samples to apply reextraction.

min_mz_delta

adjust min and max m/z so that the difference is at least this value.

min_rt_delta

adjust min and max rt so that the difference is at least this value.

verbose

flag for explicit printing

debug

flag for debugging

Value

intensities_table Each row indicates quant values for a single peak (an annotated feature in a single sample.) Both aligned RT coordinates and sample RT coordinates are given.

sample:

sample name

sampleId:

ID number for sample in mzrolldb files

groupId:

group ID number from curated_mzrolldb file

compoundName:

String representation of annotated feature

intensity:

curated_mzrollDB-recorded intensity (peakAreaTop), or reextracted intensity

is_identified:

if TRUE, using quant from curated_mzrolldb. Else, reextract from sample file

mzmin:

minimum m/z associated with peak

mzmax:

maximum m/z associated with peak

aligned_rtmin:

minimum RT value associated with peak, in aligned RT space. Relies on rt_update_key table in full_mzrolldb

aligned_rtmax:

maximum RT value associated with peak, in aligned RT space. Relies on rt_update_key table in full_mzrolldb

sample_rtmin:

minimum RT value associated with peak, in mzML/mzXML sample space.

sample_rtmax:

maximum RT value associated with peak, in mzML/mzXML sample space.