Read a mass spec file with mzR

This function wraps and reformats openMSfile converting its output to a list of tibbles. It can open any of the file formats of openMSfile (mzXML, mzML, netCDF, mzData) as well as g-zipped versions of these formats.

tidy_mzR_from_msfile(ms_file_path, mzroll_db_path = NULL, msLevels = NULL)

Arguments

ms_file_path

Path to an mzXML, mzML, netCDF or mzData file

mzroll_db_path

path to a .mzrollDB SQLite file.

msLevels

Filter for specific MS levels (if provided, this must be a vector of integers)

Value

tidy_mzR: a list containing two tibbles:

header: summary of scan; one scan per row

acquisitionNum: scan #

polarity: positive (1) or negative (0) mode

other scan attributes

scan_data: abundances of ions in each scan; contains three variables: mz, ic, and scan

Details

If a mzroll_db_path is provided, then use this file's rt_update_key table (if it exists) to update retention times.

See also